requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.11.2
bidict==0.22.1
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
ctypescrypto==0.5
discord==2.3.2
discord-py-interactions==5.11.0
discord-typings==0.7.0
discord.py==2.3.2
emoji==2.10.1
Flask==2.2.3
Flask-Cors==3.0.10
Flask-SocketIO==5.3.3
fleep==1.0.1
frozenlist==1.4.1
idna==3.4
inputimeout==1.0.4
interactions.py==5.11.0
itsdangerous==2.1.2
Jinja2==3.1.2
libarchive-c==5.0
MarkupSafe==2.1.2
multidict==6.0.5
netifaces==0.11.0
packaging==23.1
Pillow==10.0.0
prettytable==3.9.0
pyaes==1.6.1
pyasn1==0.5.1
python-engineio==4.4.1
python-socketio==5.8.0
pywin32==306
requests==2.28.2
rsa==4.9
simplejson==3.19.1
six==1.16.0
soupsieve==2.4
Telethon==1.33.1
tomli==2.0.1
typing_extensions==4.9.0
urllib3==1.26.15
wcwidth==0.2.6
websockets==12.0
Werkzeug==2.2.3
yarl==1.9.4
